show modal error on query isError
