--- src/components/modals/Error.moda.tsx
@@ -1,44 +1,0 @@
-import { useState } from 'react';
-import { Box, Modal, Typography } from '@mui/material';
-
-export const useModal = () => {
-  const [open, setOpen] = useState(true);
-
-  const style = {
-    position: 'absolute' as 'absolute',
-    top: '50%',
-    left: '50%',
-    transform: 'translate(-50%, -50%)',
-    width: 400,
-    bgcolor: 'background.paper',
-    border: '2px solid #000',
-    boxShadow: 24,
-    p: 4,
-  };
-
-  const openModal = () => {
-    setOpen(true);
-  };
-
-  const closeModal = () => {
-    setOpen(false);
-  };
-
-  const ModalComponent = () => (
-    <Modal open={open} onClose={closeModal}>
-      <Box sx={style}>
-        <Typography id="modal-modal-title" variant="h6" component="h2">
-          Server fetching failed
-        </Typography>
-        <Typography id="modal-modal-description" sx={{ mt: 2 }}>
-          Duis mollis, est non commodo luctus, nisi erat porttitor ligula.
-        </Typography>
-      </Box>
-    </Modal>
-  );
-
-  return {
-    openModal,
-    ModalComponent,
-  };
-};
